+++ src/main/java/ru/yandex/practicum/filmorate/storage/GenreStorage.java
@@ -2,5 +2,5 @@ package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Genre;
 
-public interface GenreStorage extends Storage<Genre, Integer> {
+public interface GenreStorage extends Storage<Genre, Long> {
 }
